\name{PPLasso-package}
\alias{PPLasso-package}
\alias{PPLasso}
\docType{package}
\title{
\packageTitle{PPLasso}
}
\description{
\packageDescription{PPLasso}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PPLasso}
\packageIndices{PPLasso}
This package provide usufull tool for the identification of prognostics and predictive biomarkers.
}
\author{
\packageAuthor{PPLasso}

Maintainer: \packageMaintainer{PPLasso}
}
\references{
W. Zhu, C. Levy-Leduc, N. Ternes. "A variable selection approach for highly correlated predictors in 
high-dimensional genomic data". (2020)
}
